Payload のインターフェースをやめた
