--- src/server/packages/User/Domain/Services/Jwt/Payload.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace User\Domain\Services\Jwt;
-
-interface Payload
-{
-    /**
-     * @return array<mixed>
-     */
-    public function toArray(): array;
-}
